Fix Factory Artifact UI Power Trackers
