--- Slider/Assets/Scripts/UI/UIPowerTracker.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b15ea08e66a1fcc4e986c54379b19f58
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
